[fix] Merge visit count fix from David
