+++ public/styles/main.css
@@ -38,6 +38,10 @@ label, button, a {
     margin-left: 2px;
 }
 
+a.menu-item, a.navbar-brand.logo-nav {
+    margin-top: 3px;
+}
+
 body,
 html {
     width: 100%;
@@ -223,9 +227,6 @@ input {
     -webkit-box-shadow: 0px 1px 4px rgba(0,0,0, 0.10);
     text-align: center
 }
-
-
-
 
 /* DISPLAY PAGE */
